refactor: Extract getKubernetesClient and resolveTargetClusters to client.go
