+++ cmd/opscart-scan/main.go
@@ -13,8 +13,6 @@ import (
 	"github.com/opscart/opscart-k8s-watcher/pkg/report"
 	"github.com/opscart/opscart-k8s-watcher/pkg/scanner"
 	"github.com/spf13/cobra"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/clientcmd"
 )
 
 var (
@@ -43,9 +41,9 @@ var (
 	wasteFormat string // waste command: output format (cli, html)
 
 	// NEW v0.7 flags (cloud costs)
-	region       string // cloud-costs: Azure region for pricing lookup
-	autoPrice    bool   // cloud-costs: auto-detect pricing from node labels
-	costFormat   string // cloud-costs: output format (table|json|html)
+	region     string // cloud-costs: Azure region for pricing lookup
+	autoPrice  bool   // cloud-costs: auto-detect pricing from node labels
+	costFormat string // cloud-costs: output format (table|json|html)
 )
 
 func main() {
@@ -709,7 +707,7 @@ Examples:
 	cloudCostsCmd.Flags().BoolVar(&allClustersFlag, "all-clusters", false, "Scan all configured clusters")
 	cloudCostsCmd.Flags().StringVar(&clusterGroupFlag, "cluster-group", "", "Scan all clusters in a group")
 	cloudCostsCmd.Flags().BoolVar(&showScenarios, "scenarios", true, "Show optimization scenarios")
-	
+
 	// Add all commands
 	rootCmd.AddCommand(configCmd)
 	rootCmd.AddCommand(emergencyCmd)
@@ -730,74 +728,6 @@ Examples:
 		os.Exit(1)
 	}
 }
-
-// ================================================================
-// Helper: resolveTargetClusters
-// ================================================================
-// Determines WHAT to scan based on flags
-func resolveTargetClusters() ([]config.ClusterConfig, bool, error) {
-	isCompare := len(compareFlag) > 0
-
-	// Case 1: --compare flag (exactly 2 clusters)
-	if isCompare {
-		if len(compareFlag) != 2 {
-			return nil, false, fmt.Errorf("--compare requires exactly 2 cluster names")
-		}
-		cfg, err := config.LoadConfig()
-		if err != nil {
-			return nil, false, err
-		}
-		a, err := cfg.GetClusterByName(compareFlag[0])
-		if err != nil {
-			return nil, false, err
-		}
-		b, err := cfg.GetClusterByName(compareFlag[1])
-		if err != nil {
-			return nil, false, err
-		}
-		return []config.ClusterConfig{*a, *b}, true, nil
-	}
-
-	// Case 2: --all-clusters flag
-	if allClustersFlag {
-		cfg, err := config.LoadConfig()
-		if err != nil {
-			return nil, false, err
-		}
-		clusters := cfg.GetAllClusters()
-		if len(clusters) == 0 {
-			return nil, false, fmt.Errorf("no clusters in config. Run: ./opscart-scan config init")
-		}
-		return clusters, false, nil
-	}
-
-	// Case 3: --cluster-group flag
-	if clusterGroupFlag != "" {
-		cfg, err := config.LoadConfig()
-		if err != nil {
-			return nil, false, err
-		}
-		clusters, err := cfg.GetClustersByGroup(clusterGroupFlag)
-		if err != nil {
-			return nil, false, err
-		}
-		return clusters, false, nil
-	}
-
-	// Case 4: --cluster flag (single cluster — existing behavior)
-	if cluster != "" {
-		return []config.ClusterConfig{
-			{Name: cluster, Context: cluster, Group: "single"},
-		}, false, nil
-	}
-
-	// Case 5: nothing specified
-	return nil, false, fmt.Errorf("specify a target:\n  --cluster <name>           Single cluster\n  --all-clusters             All configured clusters\n  --cluster-group <group>    All clusters in a group\n  --compare <a> <b>          Compare two clusters")
-}
-
-// ================================================================
-// Extracted scan functions (existing logic moved to functions)
-// ================================================================
 
 func runEmergencyScan(clusterContext string) error {
 	fmt.Printf("\n🔍 Cluster: %s\n", clusterContext)
@@ -971,13 +901,13 @@ func runCloudCostsScan(clusterContext string) error {
 
 	// ── Step 8: Build CloudCostReport ────────────────────────────────────
 	report := &models.CloudCostReport{
-		Timestamp:     time.Now(),
-		ClusterName:   clusterContext,
-		Region:        detectedRegion,
-		Provider:      "azure",
-		NodePoolCosts: poolCosts,
-		TotalNodeCost: totalNodeCost,
-		NamespaceCosts: nsCosts,
+		Timestamp:        time.Now(),
+		ClusterName:      clusterContext,
+		Region:           detectedRegion,
+		Provider:         "azure",
+		NodePoolCosts:    poolCosts,
+		TotalNodeCost:    totalNodeCost,
+		NamespaceCosts:   nsCosts,
 		TotalMonthlyCost: totalNodeCost,
 		TotalAnnualCost:  totalNodeCost * 12,
 		CostBreakdown: models.CostBreakdown{
@@ -1602,31 +1532,6 @@ func extractResourceNames(issues []models.SecurityIssue, issueType string, limit
 	}
 
 	return resources
-}
-
-// ================================================================
-// Existing helper (unchanged)
-// ================================================================
-
-// getKubernetesClient creates a Kubernetes clientset for the given cluster
-func getKubernetesClient(clusterContext string) (*kubernetes.Clientset, error) {
-	loadingRules := clientcmd.NewDefaultClientConfigLoadingRules()
-	configOverrides := &clientcmd.ConfigOverrides{
-		CurrentContext: clusterContext,
-	}
-
-	kubeConfig := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(loadingRules, configOverrides)
-	config, err := kubeConfig.ClientConfig()
-	if err != nil {
-		return nil, fmt.Errorf("failed to load kubeconfig: %w", err)
-	}
-
-	clientset, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create kubernetes client: %w", err)
-	}
-
-	return clientset, nil
 }
 
 func runNetworkScan(clusterContext string) error {
